refactor: streamline folder nesting
